Finished all class work
Going to update with comments shortly

+++ Intro To Programming/Chapter5/src/ProgrammingExcercise/Factorial.java
@@ -7,8 +7,8 @@ public class Factorial {
 	public static void main(String[] args) {
 		System.out.println("Enter a positive integer");
 			Scanner s = new Scanner(System.in);
-			int x = s.nextInt();
-			int n = 1;
+			short x = s.nextShort();
+			long n = 1;
 			
 			for(int i = 1 ; i <= x ; i++) {
 				n = n * i;
